meson: also build Doxygen docs
We need to generate the Doxyfile from a
template to deal with the builddir vs. srcdir
problematique which is required by Meson but
not catered for by the make-based build (and
the cmake build doesn't build docs at all).

Also, we need to do two separate doxygen runs,
one for the html output and one for the man pages.
This is because the html docs and the man pages
need to be installed into different locations.
Also, we only want to install two of the generated
man pages and ignore the rest.